requirements.txt
# AI DJ Co-Pilot - Dependencies

# Audio Analysis
essentia-tensorflow>=2.1
librosa>=0.10.0
madmom>=0.17.0

# Stem Separation  
demucs>=4.0.0

# Audio I/O & Processing
sounddevice>=0.4.6
pydub>=0.25.1
numpy>=1.24.0
scipy>=1.11.0

# ML/AI
torch>=2.0.0
transformers>=4.30.0
stable-baselines3>=2.0.0

# Backend
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
redis>=4.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# MIDI
mido>=1.3.0
python-rtmidi>=1.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
